:fire: Move and rename base32 files
