+++ tests/TestBase32.cpp
@@ -16,7 +16,7 @@
  */
 
 #include "TestBase32.h"
-#include "tools/base32.h"
+#include "core/Base32.h"
 #include <QTest>
 
 QTEST_GUILESS_MAIN(TestBase32)
